requirements.txt
fastapi
uvicorn
pydantic
openai
httpx
pyyaml
starlette
